fix(lint): clear 131 of 153 clippy/rustc warnings; allow the rest with rationale
CI's clippy step had never run to completion here, so warnings had accumulated
unseen. `-D warnings` also promotes rustc's dead_code/unused lints, so the real
starting count was 153, not the 5 the first (fail-fast) run suggested.

Auto-fixed via `cargo clippy --fix`: 26 files, zero failures to apply.

Fixed by hand where the rewrite is behaviour-identical and I could verify the
site: unnecessary_sort_by (x2), wildcard_in_or_patterns (x2), manual_strip (x4
sites), field_reassign_with_default, format_in_format_args, needless_range_loop
(x2 -- only the ones writing a single buffer), unnecessary_fallible_conversions,
and a needless_borrows/collapsible_if pass.

Two of those are more than style. `manual_strip` replaced `line[1..]` with
`strip_prefix`, removing a latent panic: the old form slices by BYTE, so a
non-ASCII FASTA header would have panicked. And the fallible-conversion fix
deleted error handling for a `UnsafeBytes` -> `SerializedBytes` conversion that
cannot fail.

Allowed rather than "fixed", each with its reason in-place, because the lint was
asking for a change that would be wrong here:

  - collapsible_match (9 integrity zomes, crate-level): every instance is the
    standard HDK validation-dispatch shape. Collapsing forces a separate
    catch-all for the remaining OpUpdate variants and restructures the exact
    RegisterUpdate path the author-binding work depends on.
  - needless_range_loop in health-zkp/circuits/range_proof.rs: clippy's suggested
    `cols.iter_mut().enumerate()` iterates the COLUMN vector while `i` indexes the
    BIT position across four columns plus an accumulator. Applying it would have
    been a real bug in circuit constraint generation.
  - needless_range_loop in health-fl/hdc_encrypted.rs: the index spans `result`
    AND every contribution's `bits` array; not a local rewrite.
  - if_same_then_else + nonminimal_bool in the HLA match_score functions: the arms
    encode distinct allele pairings in transplant-compatibility logic.
  - vec_init_then_push in hla.rs sample_typing: every element is an rng draw, so
    evaluation order determines the sampled typing.
  - large_enum_variant on two EntryTypes enums, matching the convention already
    used by 5+ sibling zomes.
  - too_many_arguments / type_complexity, matching sdk/src/lib.rs's existing
    crate-level allows.

Dead code marked with #[allow(dead_code)] and a review note rather than deleted:
removing a field from a clinical or provenance type can change serialization, and
that is a product decision.

22 warnings remain, deliberately untouched -- mostly never-used functions and
structs whose removal is a product call. One is NOT dead code and wants a real
fix: fhir_bridge/coordinator/src/lib.rs:215 assigns `patient_fhir_id` from
`get_resource_id(resource)` and never reads it, so FHIR bundle processing extracts
a patient identifier and discards it. Silencing that with an allow would hide a
probable functional gap.

Verified: cargo fmt --check clean, cargo check --workspace --lib clean.

+++ zomes/trials/coordinator/src/lib.rs
@@ -135,8 +135,8 @@ fn track_visit_data_internal(
         data_points_count,
         collected_at: visit
             .actual_date
-            .map(|ts| ts.as_micros() as i64)
-            .unwrap_or_else(|| sys_time().map(|t| t.as_micros() as i64).unwrap_or(0)),
+            .map(|ts| ts.as_micros())
+            .unwrap_or_else(|| sys_time().map(|t| t.as_micros()).unwrap_or(0)),
     };
 
     // Call dividends zome to record usage
